Finished chapter 5 of the Games from Scratch - C++ Edition
We implemented a game object manager that manages memory for game
objects, and allows us to abstract away the rendering so we don't have
to manually render objects.

This should eliminate memory leaks (kind of, I think there are still
leaks in the menu and splash screen), and although a little overkill for
a pong clone, it is a useful technique for larger games.