Finished part 6 of the Games from Scratch tutorial
In this part, we implemented the ball we're going to play with, and
movement for the paddle. I also marked locations where I believe
refactoring would be good, and I also deviated a little from the
tutorial when it game to the paddle's behavior and bouncing from the
walls.

The original tutorial had a bug where if the paddle's velocity was too
big, if you hit the edge of the game it would slowly move out of the
window. I fixed that issue, and you can see what my fix was. I didn't
document the fix in the literati source code, as I believe the actual
code is clear enough.