Finished part 9 of Game From Scratch tutorial
Implemented an AI paddle that will follow the ball around on the x axis,
and removed the code that allows the ball to bounce off the top of the
screen.

All we need to do is to implement a scoring system and a win/lose
sequence, and we will have a very buggy, very ugly game.

Given that this is where the tutorials end on the Game from Scratch
website, I will continue alone. Next I will implement scoring and a UI
for score keeping, and I will implement winning/losing sequences.

After that, I will fix the bugs with ball collisions and AI, and tweak
the controls so that it doesn't feel so sluggish.