Refactored the game a little and fixed bugs with ball interaction
I have refactored common paddle functionality into an abstract base
class, Paddle. I have implemented paddle/ball interaction within a loop,
so that both the AI paddle and the player paddle have the same
restrictions and abilities.

I have fixed a few bugs in the interaction of the ball and the paddles,
and added some new behavior. Now when you strike the ball with the side
of your paddle, it will deflect off screen and you will lose, instead of
it moving to the top of the paddle and reflecting back.

Now there are 2 more known issues to fix before I move into a scoring
system - one with the AI jitter, and one with the ball speed coming off
of a paddle hit. Sometimes you can hit the ball multiple times and it
will vastly increase it's speed.