Fixed jittery AI paddle
Gave the AI paddle the ability to predict where the ball will be,
without taking into account bouncing from walls. Also the AI paddle
tracks the ball's position only when the ball is coming towards it. When
the ball is heading towards the user paddle, the AI paddle tracks it
within a threshold.