Fixed AI, sounds and bouncing
Added refinements to the AI logic that will mean the AI can be beaten.
This required slowing the paddles down so that they're not fast enough
to beat the ball to the other side of the stage, but it didn't
negatively affect balance.

I also fixed and simplified the bouncing logic so that the ball bounces
how you'd expect it to.

I also added sound for when the ball goes off the stage.