Finished game part of Pang - just need polish
I have created game ending flows with winning and losing screens, and
the game can now determine when the winning score (12 right now) has
been reached, and whether the player or the AI won.

I have also moved the score keeping off of the ball (which never made
sense) and put it in the game.

I have also fixed a few bugs with bouncing and waiting on the menu too
long causing the player to lose the first point.