Finished Pang, the Pong clone
I have cut a few features, and made some final polishing changes, and am
calling this game done.

I changed the splash screen and the menu to suit the new look. I removed
the splash music because it didn't match with the feel of the game.

I scrapped a new splash graphic and new splash music, on account of not
having the skill to do either. I also scrapped the AI paddle idle polish
because I'm quickly running out of steam on this little game.

It's done!